# src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.unknown_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"Record_string.any_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"DidDocument": {
				"$ref": "#/components/schemas/Record_string.any_"
			},
...
			"DidNymTransaction": {
				"properties": {
					"did": {
						"type": "string"
					},
					"nymRequest": {
						"type": "string"
					}
				},
				"required": [
					"did",
					"nymRequest"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EndorserTransaction": {
				"properties": {
					"transaction": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Record_string.unknown_"
							}
						]
					},
					"endorserDid": {
						"type": "string"
					}
				},
				"required": [
					"transaction",
					"endorserDid"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Version": {
				"type": "string",
				"example": "1.0.0"
			},
			"CreateSchemaInput": {
				"properties": {
					"issuerId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"version": {
						"$ref": "#/components/schemas/Version"
					},
					"attributes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"endorse": {
						"type": "boolean"
					},
					"endorserDid": {
						"type": "string"
					}
				},
				"required": [
					"issuerId",
					"name",
					"version",
					"attributes"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"WriteTransaction": {
				"properties": {
					"endorsedTransaction": {
						"type": "string"
					},
					"endorserDid": {
						"type": "string"
					},
					"schema": {
						"properties": {
							"attributes": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"version": {
								"$ref": "#/components/schemas/Version"
							},
							"name": {
								"type": "string"
							},
							"issuerId": {
								"type": "string"
							}
						},
						"required": [
							"attributes",
							"version",
							"name",
							"issuerId"
						],
						"type": "object"
					},
					"credentialDefinition": {
						"properties": {
							"type": {
								"type": "string"
							},
							"value": {},
							"tag": {
								"type": "string"
							},
							"issuerId": {
								"type": "string"
							},
							"schemaId": {
								"type": "string"
							}
						},
						"required": [
							"type",
							"value",
							"tag",
							"issuerId",
							"schemaId"
						],
						"type": "object"
					}
				},
				"required": [
					"endorsedTransaction"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SchemaId": {
				"type": "string",
				"example": "WgWxqztrNooG92RXvxSTWv:2:schema_name:1.0"
			},
...
			"DidRegistrationSecretOptions": {
				"$ref": "#/components/schemas/Record_string.unknown_"
			},
			"DidOperationStateWait": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"wait"
						],
						"nullable": false
					},
					"did": {
						"type": "string"
					},
					"secret": {
						"$ref": "#/components/schemas/DidRegistrationSecretOptions"
					},
					"didDocument": {
						"$ref": "#/components/schemas/DidDocument"
					}
				},
				"required": [
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DidOperationStateActionBase": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"action"
						],
						"nullable": false
					},
					"action": {
						"type": "string"
					},
					"did": {
						"type": "string"
					},
					"secret": {
						"$ref": "#/components/schemas/DidRegistrationSecretOptions"
					},
					"didDocument": {
						"$ref": "#/components/schemas/DidDocument"
					}
				},
				"required": [
					"state",
					"action"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DidOperationStateFinished": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"finished"
						],
						"nullable": false
					},
					"did": {
						"type": "string"
					},
					"secret": {
						"$ref": "#/components/schemas/DidRegistrationSecretOptions"
					},
					"didDocument": {
						"$ref": "#/components/schemas/DidDocument"
					}
				},
				"required": [
					"state",
					"did",
					"didDocument"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DidOperationStateFailed": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"failed"
						],
						"nullable": false
					},
					"did": {
						"type": "string"
					},
					"secret": {
						"$ref": "#/components/schemas/DidRegistrationSecretOptions"
					},
					"didDocument": {
						"$ref": "#/components/schemas/DidDocument"
					},
					"reason": {
						"type": "string"
					}
				},
				"required": [
					"state",
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DidRegistrationMetadata": {
				"$ref": "#/components/schemas/Record_string.unknown_"
			},
			"DidResolutionMetadata": {
				"properties": {
					"contentType": {
						"type": "string"
					},
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									"invalidDid",
									"notFound",
									"representationNotSupported",
									"unsupportedDidMethod"
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"servedFromCache": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DidCreateResult_DidOperationStateActionBase_": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"didState": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/DidOperationStateWait"
							},
							{
								"$ref": "#/components/schemas/DidOperationStateActionBase"
							},
							{
								"$ref": "#/components/schemas/DidOperationStateFinished"
							},
							{
								"$ref": "#/components/schemas/DidOperationStateFailed"
							}
						]
					},
					"didRegistrationMetadata": {
						"$ref": "#/components/schemas/DidRegistrationMetadata"
					},
					"didDocumentMetadata": {
						"$ref": "#/components/schemas/DidResolutionMetadata"
					}
				},
				"required": [
					"didState",
					"didRegistrationMetadata",
					"didDocumentMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AnonCredsSchema": {
				"properties": {
					"issuerId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"version": {
						"type": "string"
					},
					"attrNames": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"issuerId",
					"name",
					"version",
					"attrNames"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsResolutionMetadata": {
				"properties": {
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									"invalid",
									"notFound",
									"unsupportedAnonCredsMethod"
								]
							}
						]
					},
					"message": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CredoExtensible": {
				"$ref": "#/components/schemas/Record_string.unknown_"
			},
			"GetSchemaReturn": {
				"properties": {
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"schemaId": {
						"type": "string"
					},
					"resolutionMetadata": {
						"$ref": "#/components/schemas/AnonCredsResolutionMetadata"
					},
					"schemaMetadata": {
						"$ref": "#/components/schemas/CredoExtensible"
					}
				},
				"required": [
					"schemaId",
					"resolutionMetadata",
					"schemaMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterSchemaReturnStateWait": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"wait"
						],
						"nullable": false
					},
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"schemaId": {
						"type": "string"
					}
				},
				"required": [
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterSchemaReturnStateAction": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"action"
						],
						"nullable": false
					},
					"action": {
						"type": "string"
					},
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"schemaId": {
						"type": "string"
					}
				},
				"required": [
					"state",
					"action",
					"schema",
					"schemaId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterSchemaReturnStateFinished": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"finished"
						],
						"nullable": false
					},
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"schemaId": {
						"type": "string"
					}
				},
				"required": [
					"state",
					"schema",
					"schemaId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterSchemaReturnStateFailed": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"failed"
						],
						"nullable": false
					},
					"reason": {
						"type": "string"
					},
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"schemaId": {
						"type": "string"
					}
				},
				"required": [
					"state",
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterSchemaReturn": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"schemaState": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/RegisterSchemaReturnStateWait"
							},
							{
								"$ref": "#/components/schemas/RegisterSchemaReturnStateAction"
							},
							{
								"$ref": "#/components/schemas/RegisterSchemaReturnStateFinished"
							},
							{
								"$ref": "#/components/schemas/RegisterSchemaReturnStateFailed"
							}
						]
					},
					"schemaMetadata": {
						"$ref": "#/components/schemas/CredoExtensible"
					},
					"registrationMetadata": {
						"$ref": "#/components/schemas/CredoExtensible"
					}
				},
				"required": [
					"schemaState",
					"schemaMetadata",
					"registrationMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "credo-controller",
		"version": "0.9.4",
		"description": "Rest endpoint wrapper for using your agent over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/polygon/create-keys": {
			"post": {
				"operationId": "CreateKeyPair",
				"responses": {
					"200": {
						"description": "Secp256k1KeyPair",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"address": {
											"type": "string"
										},
										"publicKeyBase58": {
											"type": "string"
										},
										"privateKey": {
											"type": "string"
										}
									},
									"required": [
										"address",
										"publicKeyBase58",
										"privateKey"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create Secp256k1 key pair for polygon DID",
				"tags": [
					"Polygon"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": []
			}
		},
		"/polygon/create-schema": {
			"post": {
				"operationId": "CreateSchema",
				"responses": {
					"200": {
						"description": "Schema JSON",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Create polygon based W3C schema",
				"tags": [
					"Polygon"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"schema": {
										"properties": {},
										"additionalProperties": {},
										"type": "object"
									},
									"schemaName": {
										"type": "string"
									},
									"did": {
										"type": "string"
									}
								},
								"required": [
									"schema",
									"schemaName",
									"did"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/polygon/estimate-transaction": {
			"post": {
				"operationId": "EstimateTransaction",
				"responses": {
					"200": {
						"description": "Transaction Object",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Estimate transaction",
				"tags": [
					"Polygon"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"transaction": {},
									"operation": {}
								},
								"required": [
									"transaction",
									"operation"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/polygon/{did}/{schemaId}": {
			"get": {
				"operationId": "GetSchemaById",
				"responses": {
					"200": {
						"description": "Schema Object",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Fetch schema details",
				"tags": [
					"Polygon"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "schemaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/multi-tenancy/transactions/set-endorser-role/{tenantId}": {
			"post": {
				"operationId": "DidNymTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"MultiTenancy"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidNymTransaction"
							}
						}
					}
				}
			}
		},
		"/multi-tenancy/transactions/endorse/{tenantId}": {
			"post": {
				"operationId": "EndorserTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"signedTransaction": {}
									},
									"required": [
										"signedTransaction"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"MultiTenancy"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EndorserTransaction"
							}
						}
					}
				}
			}
		},
...
		"/multi-tenancy/schema/{tenantId}": {
			"post": {
				"operationId": "CreateSchema",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"MultiTenancy"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSchemaInput"
							}
						}
					}
				}
			}
		},
...
		"/multi-tenancy/transactions/write/{tenantId}": {
			"post": {
				"operationId": "WriteSchemaAndCredDefOnLedger",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"MultiTenancy"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WriteTransaction"
							}
						}
					}
				}
			}
		},
		"/multi-tenancy/schema/{schemaId}/{tenantId}": {
			"get": {
				"operationId": "GetSchemaById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"MultiTenancy"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "schemaId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SchemaId"
						}
					},
					{
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/transactions/endorse": {
			"post": {
				"operationId": "EndorserTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"signedTransaction": {}
									},
									"required": [
										"signedTransaction"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"EndorserTransaction"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EndorserTransaction"
							}
						}
					}
				}
			}
		},
		"/transactions/set-endorser-role": {
			"post": {
				"operationId": "DidNymTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidCreateResult_DidOperationStateActionBase_"
								}
							}
						}
					}
				},
				"tags": [
					"EndorserTransaction"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidNymTransaction"
							}
						}
					}
				}
			}
		},
		"/transactions/write": {
			"post": {
				"operationId": "WriteSchemaAndCredDefOnLedger",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"EndorserTransaction"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WriteTransaction"
							}
						}
					}
				}
			}
		},
...
		"/schemas/{schemaId}": {
			"get": {
				"operationId": "GetSchemaById",
				"responses": {
					"200": {
						"description": "get schema by Id",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetSchemaReturn"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ver": "1.0",
											"id": "WgWxqztrNooG92RXvxSTWv:2:schema_name:1.0",
											"name": "schema",
											"version": "1.0",
											"attrNames": [
												"string"
											],
											"seqNo": 351936
										}
									}
								}
							}
						}
					}
				},
				"description": "Get schema by schemaId",
				"tags": [
					"Schemas"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "schemaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/schemas": {
			"post": {
				"operationId": "CreateSchema",
				"responses": {
					"200": {
						"description": "get schema",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/RegisterSchemaReturn"
										},
										{
											"$ref": "#/components/schemas/RegisterSchemaReturnStateFinished"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"state": "finished",
											"schema": {
												"issuerId": "did:indy:bcovrin:testnet:LRCUFcizUL74AGgLqdJHK7",
												"name": "Test Schema",
												"version": "1.0.0",
												"attrNames": [
													"Name",
													"Age"
												]
											},
											"schemaId": "LRCUFcizUL74AGgLqdJHK7:2:Test Schema:1.0.0"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create schema",
				"tags": [
					"Schemas"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSchemaInput"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}